Use setproctitle if available to show state
setproctitle() changes the process name, particularly in top/ps output
on posix-y systems.  This is very useful to get some transparency on
what's going on, especially if a test starts taking a very long time or
straight up hangs.

This is strictly a "bonus goodie", if setproctitle is not available,
it's not used, and any errors are ignored silently.

Signed-off-by: David Lamparter <equinox@diac24.net>