Make node ID modification more stable.
We force test collection on the controller node, as we run some global
setup there before starting the workers. That setup also depends on the
test node IDs, but currently, pytest-xdist only modifies the test node IDs
on the workers, that've been started and where `setup_config` has been called.
I don't see a good reason for this, and for creating a separate config entry,
so let's just check the right condition in the pytest hook.
That way, the hook will always do the same node ID transformation, regardless
of pytest-xdist setup.