Loadgroup: Don't modify test nodeids any longer.
The loadgroup scheduler used to modify the nodeids of
tests that were placed manually in a specific scope.
This was used as a side-channel from test collection (on the pytest-running process)
to the worker process. Instead, make that side-channel explicit by passing
a parameter around.
That's pretty ugly, but no less ugly than modifying and parsing (!)
node IDs later on.